redo function optimizer PR
