+++ tests/test_get_helper_code.py
@@ -3,10 +3,13 @@ from argparse import Namespace
 from pathlib import Path
 
 import pytest
+
 from codeflash.discovery.functions_to_optimize import FunctionToOptimize
 from codeflash.either import is_successful
 from codeflash.models.models import FunctionParent
+from codeflash.optimization.function_optimizer import FunctionOptimizer
 from codeflash.optimization.optimizer import Optimizer
+from codeflash.verification.verification_utils import TestConfig
 
 
 class HelperClass:
@@ -31,6 +34,7 @@ def test_get_outside_method_helper() -> None:
             experiment_id=None,
         )
     )
+
     function_to_optimize = FunctionToOptimize(
         function_name="OptimizeMe", file_path=file_path, parents=[], starting_line=None, ending_line=None
     )
@@ -51,7 +55,7 @@ _KEY_T = TypeVar("_KEY_T")
 _STORE_T = TypeVar("_STORE_T")
 class AbstractCacheBackend(CacheBackend, Protocol[_KEY_T, _STORE_T]):
     """Interface for cache backends used by the persistent cache decorator."""
-    
+
     def __init__(self) -> None: ...
 
     def hash_key(
@@ -213,17 +217,6 @@ class _PersistentCache(Generic[_P, _R, _CacheBackendT]):
         f.write(code)
         f.flush()
         file_path = Path(f.name).resolve()
-        opt = Optimizer(
-            Namespace(
-                project_root=file_path.parent.resolve(),
-                disable_telemetry=True,
-                tests_root="tests",
-                test_framework="pytest",
-                pytest_cmd="pytest",
-                experiment_id=None,
-                test_project_root=Path().resolve(),
-            )
-        )
         function_to_optimize = FunctionToOptimize(
             function_name="__call__",
             file_path=file_path,
@@ -231,17 +224,25 @@ class _PersistentCache(Generic[_P, _R, _CacheBackendT]):
             starting_line=None,
             ending_line=None,
         )
+        test_config = TestConfig(
+            tests_root="tests",
+            tests_project_rootdir=Path.cwd(),
+            project_root_path=file_path.parent.resolve(),
+            test_framework="pytest",
+            pytest_cmd="pytest",
+        )
+        func_optimizer = FunctionOptimizer(function_to_optimize=function_to_optimize, test_cfg=test_config)
         with open(file_path) as f:
             original_code = f.read()
-        ctx_result = opt.get_code_optimization_context(function_to_optimize, opt.args.project_root, original_code)
+        ctx_result = func_optimizer.get_code_optimization_context()
         if not is_successful(ctx_result):
             pytest.fail()
         code_context = ctx_result.unwrap()
         assert code_context.helper_functions[0].qualified_name == "AbstractCacheBackend.get_cache_or_call"
 
         assert (
-            code_context.code_to_optimize_with_helpers
-            == '''_R = TypeVar("_R")
+                code_context.code_to_optimize_with_helpers
+                == '''_R = TypeVar("_R")
 
 class AbstractCacheBackend(CacheBackend, Protocol[_KEY_T, _STORE_T]):
     def __init__(self) -> None: ...
@@ -338,29 +339,27 @@ class _PersistentCache(Generic[_P, _R, _CacheBackendT]):
 
 def test_bubble_sort_deps() -> None:
     file_path = (Path(__file__) / ".." / ".." / "code_to_optimize" / "bubble_sort_deps.py").resolve()
-    opt = Optimizer(
-        Namespace(
-            project_root=file_path.parent.parent.resolve(),
-            disable_telemetry=True,
-            tests_root=str(file_path.parent / "tests"),
-            test_framework="pytest",
-            pytest_cmd="pytest",
-            experiment_id=None,
-            test_project_root=file_path.parent.resolve(),
-        )
-    )
+
     function_to_optimize = FunctionToOptimize(
         function_name="sorter_deps", file_path=file_path, parents=[], starting_line=None, ending_line=None
     )
+    test_config = TestConfig(
+        tests_root=str(file_path.parent / "tests"),
+        tests_project_rootdir=file_path.parent.resolve(),
+        project_root_path=file_path.parent.parent.resolve(),
+        test_framework="pytest",
+        pytest_cmd="pytest",
+    )
+    func_optimizer = FunctionOptimizer(function_to_optimize=function_to_optimize, test_cfg=test_config)
     with open(file_path) as f:
         original_code = f.read()
-    ctx_result = opt.get_code_optimization_context(function_to_optimize, opt.args.project_root, original_code)
+    ctx_result = func_optimizer.get_code_optimization_context()
     if not is_successful(ctx_result):
         pytest.fail()
     code_context = ctx_result.unwrap()
     assert (
-        code_context.code_to_optimize_with_helpers
-        == """def dep1_comparer(arr, j: int) -> bool:
+            code_context.code_to_optimize_with_helpers
+            == """def dep1_comparer(arr, j: int) -> bool:
     return arr[j] > arr[j + 1]
 
 def dep2_swap(arr, j):
@@ -378,7 +377,7 @@ def sorter_deps(arr):
     )
     assert len(code_context.helper_functions) == 2
     assert (
-        code_context.helper_functions[0].fully_qualified_name
-        == "code_to_optimize.bubble_sort_dep1_helper.dep1_comparer"
+            code_context.helper_functions[0].fully_qualified_name
+            == "code_to_optimize.bubble_sort_dep1_helper.dep1_comparer"
     )
     assert code_context.helper_functions[1].fully_qualified_name == "code_to_optimize.bubble_sort_dep2_swap.dep2_swap"
